feat(asr): move WAV handling to its own class
